+++ src/main/resources/static/css/modal.css
@@ -2,7 +2,7 @@
     position: absolute;
     width: 100%;
     height: 100%;
-    background: rgba(0, 0, 0, 0.2);
+    background: rgba(0, 0, 0, 1);
     top: 0;
     left: 0;
     display: none;
@@ -48,4 +48,10 @@
 
 #hr_divide {
     border-top: 1px solid;
+}
+
+#academyTable {
+    margin-left:auto;
+    margin-right:auto;
+    width: 340px;
 }
